fix read_range_prefix_n
Fix bug where prefix of key wasn't matched properly.
To be able to match key we need to encode the key and prefix as an erlang list.
Since sext will encode it so that we can compare the list until they differ.
I.e. there are no number of elements encoded in the beginning.
This enables us to match the prefix of the key.